--- web/src/lib/parseMoney.ts
@@ -1,8 +1,0 @@
-export const parseMoney = (cents: number) => {
-  const dollars = cents / 100;
-
-  return dollars.toLocaleString('en-US', {
-    style: 'currency',
-    currency: 'USD',
-  });
-};
