+++ web/src/pages/TransactionPage/index.tsx
@@ -3,7 +3,7 @@ import { useMemo } from 'react';
 import { TableContainer } from './style';
 import { Column, Cell } from 'react-table';
 import { Transaction } from '@Generated/graphql';
-import { parseMoney } from '@Lib/parseMoney';
+import { formatMoneyFromCentsToDollars } from '@Lib/money-utils';
 import { useGetTransactionsQuery } from '@Generated/graphql';
 import { format } from 'date-fns';
 import Table from '@Modules/transactions/Table';
@@ -50,7 +50,9 @@ function TransactionPage() {
         Header: () => <span className="align-right">Amount</span>,
         accessor: 'amount',
         Cell: ({ value }: Cell<Transaction>) => {
-          return <div className="number">{parseMoney(value)}</div>;
+          return (
+            <div className="number">{formatMoneyFromCentsToDollars(value)}</div>
+          );
         },
         Filter: SelectTypeFilter,
         disableFilters: true,
